+++ examples/03.Blink/blink.py
@@ -15,3 +15,11 @@ except KeyboardInterrupt:
 finally:
     GPIO.cleanup()
     print('GPIO cleanup completed.')
+
+#
+#(2) Импортируем модуль time для функции задержки времени.
+#(8) Константа true всегда истинна, поэтому цикл будет выполняться вечно.
+#(9) Вызываем функцию sleep из модуля time. Параметр 0.5 задаёт задержку в полсекунды.
+#(13) Отлавливаем исключение сообщающее о прерывании программы клавишами ctrl+C.
+#(16-17) Поместив внутрь блока finally функцию GPIO.cleanup(), можно быть уверенным, что пины Raspberry сбросятся в начальное состояние.
+#
