+++ examples/06.LedBrightness/brightnessControl.py
@@ -20,3 +20,11 @@ except KeyboardInterrupt:
 finally:
     GPIO.cleanup()
     print('GPIO cleanup completed.')
+
+#
+#(6) Пин 18 будет работать в режиме ШИМ на частоте 1000 герц.
+#(7) Запускаем ШИМ на пине с коэффициентом заполнения 50%, это зажжёт светодиод вполсилы.
+#(13) На каждой итерации цикла увеличиваем значение переменной на 1.
+#(14-15) Коэффициент заполнения принимает значения от 0 до 100. Если вышли за допустимые границы, обнуляем значение.
+#(16) Обновляем коэффициент заполнения новым значением dutyCycle.
+#
