+++ examples/07.LightControlPanel/function.py
@@ -1,12 +1,10 @@
 import RPi.GPIO as GPIO
 
-
 def isPressed(btn, led):
     if GPIO.input(btn) == False:
         GPIO.output(led, GPIO.HIGH)
     else:
         GPIO.output(led, GPIO.LOW)
-
 
 button1 = 3
 button2 = 4
@@ -28,3 +26,8 @@ except KeyboardInterrupt:
 finally:
     GPIO.cleanup()
     print('GPIO cleanup completed.')
+
+#
+#(3-7) Создаём функцию с именем isPressed. Говорим, что она принимает 2 параметра: пин кнопки и пин светодиода. В теле функции пишем условие if ... else ... для управления освещением.
+#(22,23) В цикле while вызываем функцию isPressed и передаём пары кнопка-светодиод.
+#
