+++ examples/09.WEBServer/simpleServer.py
@@ -1,9 +1,16 @@
 from flask import Flask
 
 app = Flask('simpleServer')
+
 @app.route('/')
 def index():
     return 'Hello, Amperka!'
 
-
 app.run(port=3000, host='0.0.0.0')
+
+#
+#(1) Импортируем библиотеку Flask.
+#(3) Создаём объект Flask и записываем его в переменную app.
+#(5-7) Вызовем функцию index, когда клиент приложения (мы) запросит корневой адрес сайта.
+#(9) Если скрипт запускается из командной строки, запустить сервер на порту 3000.
+#
