+++ examples/13.WeatherPhidget/phidget.py
@@ -1,5 +1,4 @@
-import requests
-import json
+import requests, json
 
 url = 'http://api.openweathermap.org/data/2.5/forecast'
 
@@ -15,7 +14,6 @@ data = json.loads(res.text)
 
 weather = data['list'][0]
 
-
 def pars_weather(weatherType, timeRange, measurementUnits):
     if (weatherType in weather) and (
         timeRange in weather[weatherType].keys()
@@ -29,8 +27,17 @@ def pars_weather(weatherType, timeRange, measurementUnits):
     else:
         print(weatherType, ': ', 'none')
 
-
 pars_weather('clouds', 'all', '%')
 pars_weather('rain', '3h', 'mm')
 pars_weather('snow', '3h', 'mm')
 print('temp:', weather['main']['temp'], 'C')
+
+#
+#(1) Подключаем библиотеку requests для генерации запросов и json для обработки ответов.
+#(3) В переменной url храним адрес сайта, к которому будем обращаться за информацией о погоде.
+#(6-9) В переменную payload записываем пары ключ-значение: широта, долгота, единицы измерения и ключ API. Для примера широта и долгота московского офиса Амперки 55.7440 и 37.6568.
+#(12) Производим запрос к сервису погоды (url) с заданными параметрами (payload). Ответ вернётся в формате JSON.
+#(13) Функцией json.loads() преобразуем JSON к словарю.
+#(17-28) Функция pars_weather проходится по словарю и, в зависимости от полученных параметров, выводит удобочитаемый прогноз погоды.
+#(33) Выводим значение температуры.
+#
