+++ examples/14.VKBot/vk.py
@@ -9,14 +9,12 @@ param = {
     'filter': 'unread'
 }
 
-
 def write_msg(user_id, msg, random):
     vk.method('messages.send', {
         'user_id': user_id,
         'message': msg,
         'random_id': random
     })
-
 
 try:
     while True:
@@ -33,3 +31,11 @@ except KeyboardInterrupt:
     print('Keyboard interrupt detected.')
 finally:
     print('The program was stopped.')
+
+#
+#(4) Инициализируем бота. Вместо твой_ключ следует подставить строку с ключом доступа приложения, который тебе присвоил ВКонтакте.
+#(6-9) Словарь param хранит параметры запроса: сколько присылать сообщений за раз (1), за какой период (5 секунд) и только непрочитанные.
+#(12-16) Создаём функцию отправки сообщений указанным пользователям user_id с заданным текстом msg. Уникальный идентификатор random предотвращает повторную отправку одинаковых сообщений.
+#(21) Раз в секунду отправляем запрос серверу ВКонтакте.
+#(22-29) Если есть новые сообщения – сохраняем id отправителя, текст сообщения, уникальный идентификатор и передаем их в функцию отправки сообщений.
+#
